requirements.txt
aiosmtplib
alembic
annotated-types
anyio
async-asgi-testclient
async-timeout
asyncpg
bcrypt
blinker
certifi
cffi
charset-normalizer
click
cryptography
dnspython
docker
ecdsa
email-validator
exceptiongroup
fastapi
fastapi-mail
fastapi-users
fastapi-users-db-sqlalchemy
greenlet
gunicorn
h11
httpcore
httptools
httpx
idna
iniconfig
itsdangerous
Jinja2
makefun
Mako
MarkupSafe
multidict
orjson
packaging
passlib
pluggy
pyasn1
pycparser
pydantic
pydantic-extra-types
pydantic-settings
pydantic_core
PyJWT
pytest
python-dotenv
python-jose
python-multipart
PyYAML
requests
rsa
six
sniffio
SQLAlchemy
SQLAlchemy-Utils
starlette
tomli
typing_extensions
ujson
urllib3
uvicorn
uvloop
watchfiles
websockets
pytest-cov